fix1

+++ src/application/jwt-service.ts
@@ -1,7 +1,8 @@
 import jwt from 'jsonwebtoken'
 import {UserType_Id} from "../types/user-type";
 import {ObjectId} from "mongodb";
-import {jwtAccess, jwtRefresh, userCollection} from "../db";
+import {jwtAccess, jwtRefresh, tokenCollection, userCollection} from "../db";
+import {tr} from "date-fns/locale";
 
 
 export const jwtService = {
@@ -9,12 +10,21 @@ export const jwtService = {
     async createJwt(user: UserType_Id) {
 
         const accessToken = jwt.sign({user: user._id}, jwtAccess, {expiresIn: '10s'})
-        const refreshToken = jwt.sign({user: user._id}, jwtRefresh, {expiresIn: '30d'})
+        const refreshToken = jwt.sign({user: user._id}, jwtRefresh, {expiresIn: '20s'})
 
         return {
             accessToken, refreshToken
         }
     },
+
+    // async saveToken(user: UserType_Id, refreshToken: string){
+    //     const token = await tokenCollection.findOne({user: user._id})
+    //     if(token){
+    //         token.refreshToken = refreshToken
+    //         return await token.save()
+    //     }
+    //     return await tokenCollection.insertOne({user: user._id, refreshToken})
+    // },
 
 
     async getUserByToken(token: string) {
@@ -28,34 +38,13 @@ export const jwtService = {
     },
 
 
-    async removeToken(refreshToken: string){
-        const tokenData = await userCollection.deleteOne({refreshToken})
-        return tokenData
-
-    },
-
-    async validateRefreshToken(refreshToken: string){
-        try {
-            const result: any = jwt.verify(refreshToken, jwtRefresh)
-            return new ObjectId(result.user)
-        } catch (error) {
-            return null
+    async removeToken(id: string ,refreshToken: string){
+        const token = await userCollection.findOne({_id: new ObjectId(id)})
+        if(token){
+            const tokenData = await userCollection.deleteOne({refreshToken})
+           return tokenData
         }
-    },
-
-
-    async refresh(refreshToken: string, user: UserType_Id){
-        if(!refreshToken){
-            return null
-        }
-        const userData = await this.validateRefreshToken(refreshToken)
-        if(!userData){
-            return null
-        }
-       await this.createJwt(user)
-
-        return
-
+        return true
     }
 
 }
